solution_clap_backup/requirements.txt
# Multi-Agent RCA System Dependencies

# Core LLM Framework
clap-agents>=0.3.0

# LLM Provider SDKs
openai>=1.0.0
groq>=0.4.0

# Async Support
anyio>=4.0.0
httpx>=0.27.0

# Environment Management
python-dotenv>=1.0.0

# Validation
jsonschema>=4.0.0
pydantic>=2.0.0

# Utilities
colorama>=0.4.0

# Optional: For development
# pytest>=7.0.0
# pytest-asyncio>=0.21.0

